fix(dcd): resolve around OpenseaLink and CollectNtfContext
