+++ src/components/[guild]/collect/components/CollectNftContext.tsx
@@ -1,16 +1,26 @@
+import { Link } from "@chakra-ui/next-js"
 import { HStack, Text } from "@chakra-ui/react"
+import { Icon, Img } from "@chakra-ui/react"
+import { ArrowSquareOut } from "@phosphor-icons/react"
 import {
   TransactionStatusProvider,
   useTransactionStatusContext,
 } from "components/[guild]/Requirements/components/GuildCheckout/components/TransactionStatusContext"
 import TransactionStatusModal from "components/[guild]/Requirements/components/GuildCheckout/components/TransactionStatusModal"
-import OpenseaLink from "components/[guild]/Requirements/components/GuildCheckout/components/TransactionStatusModal/components/OpenseaLink"
 import { RewardIcon } from "components/[guild]/RoleCard/components/Reward"
 import { RewardDisplay } from "components/[guild]/RoleCard/components/RewardDisplay"
 import { ContractCallFunction } from "components/[guild]/RolePlatforms/components/AddRoleRewardModal/components/AddContractCallPanel/components/CreateNftForm/hooks/useCreateNft"
-import { PropsWithChildren, createContext, useContext, useEffect } from "react"
+import {
+  PropsWithChildren,
+  ReactNode,
+  createContext,
+  useContext,
+  useEffect,
+} from "react"
 import { FormProvider, useForm, useWatch } from "react-hook-form"
 import { GuildPlatform } from "types"
+import { openseaBaseUrl } from "utils/guildCheckout/constants"
+import { useChainId } from "wagmi"
 import { Chain, Chains } from "wagmiConfig/chains"
 import useGuildRewardNftBalanceByUserId from "../hooks/useGuildRewardNftBalanceByUserId"
 import useNftDetails from "../hooks/useNftDetails"
@@ -29,7 +39,7 @@ export type CollectNftForm = {
   amount: number
 }
 
-const CollectNftContext = createContext<Props>(undefined)
+const CollectNftContext = createContext<Props>({} as Props)
 
 const CollectNftProvider = ({
   roleId,
@@ -171,5 +181,22 @@ const CollectNftProviderWrapper = ({
 )
 
 const useCollectNftContext = () => useContext(CollectNftContext)
+
+const OpenseaLink = (): ReactNode => {
+  const chainId = useChainId()
+  const { nftAddress } = useCollectNftContext()
+  const baseUrl = openseaBaseUrl[Chains[chainId]]
+
+  if (!baseUrl) return null
+  return (
+    <Text mb={6} colorScheme="gray">
+      <Link isExternal href={`${baseUrl}/${nftAddress}`}>
+        <Img src={"/requirementLogos/opensea.svg"} boxSize={"1em"} mr="1.5" />
+        View on OpenSea
+        <Icon ml={1.5} as={ArrowSquareOut} />
+      </Link>
+    </Text>
+  )
+}
 
 export { CollectNftProviderWrapper as CollectNftProvider, useCollectNftContext }
